// gravitee-apim-e2e/lib/portal-webclient-sdk/src/lib/models/Api.ts
export interface Api {
    /**
     * Unique identifier of an API.
     * @type {string}
     * @memberof Api
     */
    id: string;
    /**
     * Name of the API.
     * @type {string}
     * @memberof Api
     */
    name: string;
    /**
     * Version of the API.
     * @type {string}
     * @memberof Api
     */
    version: string;
    /**
     * Description of the API.
     * @type {string}
     * @memberof Api
     */
    description: string;
    /**
     * Whether or not the API is in draft.
     * @type {boolean}
     * @memberof Api
     */
    draft?: boolean;
    /**
     * Whether or not the API is public.
     * @type {boolean}
     * @memberof Api
     */
    _public?: boolean;
    /**
     * Whether or not the API is running.
     * @type {boolean}
     * @memberof Api
     */
    running?: boolean;
    /**
     * List of all the avaible endpoints to call the API.
     * @type {Array<string>}
     * @memberof Api
     */
    entrypoints?: Array<string>;
    /**
     * List of labels linked to this API.
     * @type {Array<string>}
     * @memberof Api
     */
    labels?: Array<string>;
    /**
     * 
     * @type {User}
     * @memberof Api
     */
    owner: User;
    /**
     * create date and time.
     * @type {Date}
     * @memberof Api
     */
    created_at?: Date;
    /**
     * Last update date and time.
     * @type {Date}
     * @memberof Api
     */
    updated_at?: Date;
    /**
     * List of categories this API belongs to.
     * @type {Array<string>}
     * @memberof Api
     */
    categories?: Array<string>;
    /**
     * 
     * @type {RatingSummary}
     * @memberof Api
     */
    rating_summary?: RatingSummary;
    /**
     * 
     * @type {ApiLinks}
     * @memberof Api
     */
    _links?: ApiLinks;
    /**
     * Only returned with (*)/apis/{apiId}*.
     * Need *include* query param to contain 'pages'.
     * 
     * The documentation pages of this API. Same as (*)/apis/{apiId}/pages*.
     * @type {Array<Page>}
     * @memberof Api
     */
    pages?: Array<Page>;
    /**
     * Only returned with (*)/apis/{apiId}*.
     * Need *include* query param to contain 'plans'.
     * 
     * The plans of this API. Same as (*)/apis/{apiId}/plans*.
     * @type {Array<Plan>}
     * @memberof Api
     */
    plans?: Array<Plan>;
}